Loading certs from all stores, and adding password support
